requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
macos-notifications==0.2.0
outcome==1.3.0.post0
pip==22.3.1
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.9.0.post0
requests==2.31.0
selenium==4.18.1
setuptools==65.5.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
python-crontab
termcolor
virtualenv